+++ src/storage/block_manager.rs
@@ -21,11 +21,16 @@ use crate::storage::writer::RecordWriter;
 pub const DEFAULT_MAX_READ_CHUNK: u64 = 1024 * 1024 * 512;
 
 /// Helper class for basic operations on blocks.
-
+///
+/// ## Notes
+///
+/// It is not thread safe and may cause data corruption if used from multiple threads,
+/// because it does not lock the block descriptor file. Use it with RwLock<BlockManager>
 pub struct BlockManager {
     path: PathBuf,
     readers: HashMap<u64, Vec<Weak<RwLock<RecordReader>>>>,
     writers: HashMap<u64, Vec<Weak<RwLock<RecordWriter>>>>,
+    file_mutex: RwLock<()>,
 }
 
 pub const DESCRIPTOR_FILE_EXT: &str = ".meta";
@@ -61,6 +66,7 @@ impl BlockManager {
             path,
             readers: HashMap::new(),
             writers: HashMap::new(),
+            file_mutex: RwLock::new(()),
         }
     }
 
@@ -68,6 +74,7 @@ impl BlockManager {
     ///
     /// # Arguments
     ///
+    /// * `block_manager` - Block manager to update state of record
     /// * `block` - Block to write to.
     /// * `record` - Record to write.
     ///
@@ -75,12 +82,12 @@ impl BlockManager {
     ///
     /// * `Ok(RecordWriter)` - weak reference to the record writer.
     pub fn begin_write(
-        &mut self,
+        block_manager: Arc<RwLock<BlockManager>>,
         block: &Block,
         record_index: usize,
     ) -> Result<Arc<RwLock<RecordWriter>>, HttpError> {
         let ts = block.begin_time.clone().unwrap();
-        let path = self.path_to_data(&ts);
+        let path = block_manager.read().unwrap().path_to_data(&ts);
 
         let content_length = block.records[record_index].end - block.records[record_index].begin;
         let writer = Arc::new(RwLock::new(RecordWriter::new(
@@ -88,21 +95,23 @@ impl BlockManager {
             block,
             record_index,
             content_length,
-            Box::new(Self::new(self.path.clone())),
+            Arc::clone(&block_manager),
         )?));
 
-        let block_id = ts_to_us(&ts);
-        match self.writers.entry(block_id) {
-            Entry::Occupied(mut e) => {
-                e.get_mut().push(Arc::downgrade(&writer));
+        {
+            let mut bm = block_manager.write().unwrap();
+            let block_id = ts_to_us(&ts);
+            match bm.writers.entry(block_id) {
+                Entry::Occupied(mut e) => {
+                    e.get_mut().push(Arc::downgrade(&writer));
+                }
+                Entry::Vacant(e) => {
+                    e.insert(vec![Arc::downgrade(&writer)]);
+                }
             }
-            Entry::Vacant(e) => {
-                e.insert(vec![Arc::downgrade(&writer)]);
-            }
+
+            bm.clean_readers_or_writers(block_id);
         }
-
-        self.clean_readers_or_writers(block_id);
-
         Ok(writer)
     }
 
@@ -232,6 +241,8 @@ pub trait ManageBlock {
 
 impl ManageBlock for BlockManager {
     fn load(&self, block_id: u64) -> Result<Block, HttpError> {
+        let _guard = self.file_mutex.read().unwrap();
+
         let proto_ts = us_to_ts(&block_id);
         let buf = std::fs::read(self.path_to_desc(&proto_ts))?;
         Block::decode(Bytes::from(buf)).map_err(|e| {
@@ -240,13 +251,16 @@ impl ManageBlock for BlockManager {
     }
 
     fn save(&self, block: &Block) -> Result<(), HttpError> {
+        let _guard = self.file_mutex.write().unwrap();
+
         let path = self.path_to_desc(block.begin_time.as_ref().unwrap());
         let mut buf = BytesMut::new();
         block.encode(&mut buf).map_err(|e| {
             HttpError::internal_server_error(&format!("Failed to encode block descriptor: {}", e))
         })?;
-        let mut file = std::fs::File::create(path)?;
+        let mut file = std::fs::File::create(path.clone())?;
         file.write_all(&buf)?;
+
         Ok(())
     }
 
@@ -264,6 +278,7 @@ impl ManageBlock for BlockManager {
     }
 
     fn finish(&self, block: &Block) -> Result<(), HttpError> {
+        let _guard = self.file_mutex.write().unwrap();
         let path = self.path_to_data(block.begin_time.as_ref().unwrap());
         let file = std::fs::OpenOptions::new()
             .read(true)
@@ -280,6 +295,8 @@ impl ManageBlock for BlockManager {
                 block_id
             )));
         }
+
+        let _guard = self.file_mutex.write().unwrap();
 
         let proto_ts = us_to_ts(&block_id);
         let path = self.path_to_data(&proto_ts);
@@ -366,7 +383,7 @@ mod tests {
 
     #[test]
     fn test_start_writing() {
-        let mut bm = setup();
+        let bm = setup();
 
         let block_id = 1;
         let mut block = bm.start(block_id, 1024).unwrap().clone();
@@ -384,8 +401,9 @@ mod tests {
 
         bm.save(&block).unwrap();
 
+        let bm_ref = Arc::new(RwLock::new(bm));
         {
-            let writer = bm.begin_write(&block, 0).unwrap();
+            let writer = BlockManager::begin_write(Arc::clone(&bm_ref), &block, 0).unwrap();
             writer
                 .write()
                 .unwrap()
@@ -393,12 +411,12 @@ mod tests {
                 .unwrap();
         }
 
-        bm.finish(&block).unwrap();
+        bm_ref.read().unwrap().finish(&block).unwrap();
     }
 
     #[test]
     fn test_remove_with_writers() {
-        let mut bm = setup();
+        let bm = setup();
         let block_id = 1;
 
         {
@@ -415,11 +433,12 @@ mod tests {
                 content_type: "".to_string(),
             });
 
-            let writer = bm.begin_write(&block, 0).unwrap();
+            let bm_ref = Arc::new(RwLock::new(bm));
+            let writer = BlockManager::begin_write(Arc::clone(&bm_ref), &block, 0).unwrap();
             assert!(!writer.read().unwrap().is_done());
 
             assert_eq!(
-                bm.remove(block_id).err(),
+                bm_ref.write().unwrap().remove(block_id).err(),
                 Some(HttpError::internal_server_error(&format!(
                     "Cannot remove block {} because it is still in use",
                     block_id